Extract the RC IR ownership and consume model into its own module
Which construct consumes which reference, which binding merely aliases one,
and how a value's references group into reference-counting units is one model
that reference-count insertion, borrow-ification, cancellation, and the
validator all share. It lived among the borrow pass's own inference, routing,
and rewrite, so a reader of any of them had to separate the two.

Pure move: no behavior change.
