+++ src/rc_ir/borrow.rs
@@ -30,55 +30,27 @@
 //! Borrow-ification leaves the caller with a retain before a borrow call and a release after it,
 //! bracketing the call with no consume between. `cancel` removes those net-zero brackets: a retain is
 //! cancellable when, on every forward path, a release un-bumps it before the value is consumed. That
-//! keeps the value `Unique` for the uniqueness analysis, the reason borrow-ification exists. The
-//! cancellation shares the object-identity (`root`) and consume-site machinery with the inference and
-//! rewrite above, so all three read the same aliasing vars.
+//! keeps the value `Unique` for the uniqueness analysis, the reason borrow-ification exists.
+//!
+//! Which construct consumes which reference, and which object a reference belongs to, is the shared
+//! model in `ownership`: inference, rewrite, and cancellation all read it, so they agree.
 
-use crate::ast::inline_llvm::LLVMGen;
 use crate::ast::name::FullName;
 use crate::ast::program::TypeEnv;
 use crate::ast::types::TypeNode;
 use crate::constants::CLOSURE_CAPTURE_IDX;
-use crate::fixstd::builtin::InlineLLVMMakeUnionBody;
 use crate::misc::{grow_stack, Map, Set};
 use crate::parse::sourcefile::Span;
 use crate::rc_ir::ast::{
     FieldPath, FuncRef, MatchArm, Ownership, OwnershipShape, RcExpr, RcExprNode, RcFunc,
     RcGlobalInit, RcProgram, RcRhs, RcState, RcVar, VarPath,
 };
-use crate::rc_ir::provenance::{boxed_leaf_paths, LeafOrigin};
+use crate::rc_ir::ownership::{
+    all_owned_units, boxed_leaves, collect_consumes, destructure_consumes, origin, rc_units,
+    rhs_consumes, truncate_to_unit, units_under, VarTable,
+};
 use crate::rc_ir::rename::fresh_rename_function;
 use std::sync::Arc;
-
-/// What binds a variable, enough to trace a leaf back to the object that produced it (its `root`).
-enum Binding {
-    /// A parameter or capture — the origin of a leaf.
-    Param,
-    /// `let x = y`: a move-bind, transparent to `root`.
-    Move(RcVar),
-    /// `let x = op(args)`: an alias when the result leaf is a pure projection of one argument,
-    /// otherwise a producer. Carries the result type to consult `result_prov`.
-    Llvm(Box<dyn LLVMGen>, Vec<RcVar>, Arc<TypeNode>),
-    /// `let x = f(args)` or a closure — an opaque producer.
-    Producer,
-    /// A `destructure` field: field `idx` of the container.
-    Field(RcVar, usize),
-    /// A `match`-arm payload: the variant tag (`None` for a catch-all), and the scrutinee.
-    Payload(RcVar, Option<usize>),
-    /// `let x = match ...`: the value each arm returns. A match binding produces nothing of its own —
-    /// it receives one of these, chosen by the path taken — so its object identity is path-dependent.
-    Join(Vec<RcVar>),
-}
-
-/// The per-function vars `origin` and the consume walk need: how each local variable is bound, which
-/// closure value targets which function, the function's own parameters, and every variable's type
-/// (so a leaf that roots at any variable can be clamped to its reference-counting unit).
-struct VarTable {
-    bindings: Map<FullName, Binding>,
-    closure_targets: Map<FullName, FuncRef>,
-    param_tys: Map<FullName, Arc<TypeNode>>,
-    var_tys: Map<FullName, Arc<TypeNode>>,
-}
 
 /// The result of borrow inference: which parameter leaves are `Own` (all others are `Borrow`), keyed
 /// by the parameter variable's name and the leaf path.
@@ -122,517 +94,6 @@ fn infer_ownership(prog: &RcProgram, type_env: &TypeEnv) -> Ownerships {
     }
 
     Ownerships { own }
-}
-
-impl VarTable {
-    /// The variable table of a function: its parameters and capture as `Param` bindings, plus the `Binding` and
-    /// type of every variable bound in its body.
-    fn of(func: &RcFunc) -> VarTable {
-        let mut vars = VarTable::empty();
-        for p in func.params.iter().chain(func.capture.iter()) {
-            vars.bindings.insert(p.name.clone(), Binding::Param);
-            vars.param_tys.insert(p.name.clone(), p.ty.clone());
-            vars.var_tys.insert(p.name.clone(), p.ty.clone());
-        }
-        collect_bindings(&func.body, &mut vars);
-        vars
-    }
-
-    /// The vars of a param-less body (a global initializer).
-    fn body_only(body: &RcExprNode) -> VarTable {
-        let mut vars = VarTable::empty();
-        collect_bindings(body, &mut vars);
-        vars
-    }
-
-    fn empty() -> VarTable {
-        VarTable {
-            bindings: Map::default(),
-            closure_targets: Map::default(),
-            param_tys: Map::default(),
-            var_tys: Map::default(),
-        }
-    }
-}
-
-/// Record every local variable's `Binding` and type (and any closure value's target function) in a
-/// function body.
-fn collect_bindings(node: &RcExprNode, vars: &mut VarTable) {
-    match node.expr.as_ref() {
-        RcExpr::Ret(_) => {}
-        RcExpr::Let(x, rhs, k) => {
-            let def = match rhs {
-                RcRhs::Var(y) => Binding::Move(y.clone()),
-                RcRhs::Llvm(llvm_gen, args) => {
-                    Binding::Llvm(llvm_gen.clone(), args.clone(), x.ty.clone())
-                }
-                RcRhs::Closure(fref, _) => {
-                    vars.closure_targets.insert(x.name.clone(), fref.clone());
-                    Binding::Producer
-                }
-                RcRhs::App(..) => Binding::Producer,
-                RcRhs::Match(scrut, arms) => {
-                    let mut arm_results = vec![];
-                    for arm in arms {
-                        vars.bindings.insert(
-                            arm.payload.name.clone(),
-                            Binding::Payload(scrut.clone(), arm.tag),
-                        );
-                        vars.var_tys
-                            .insert(arm.payload.name.clone(), arm.payload.ty.clone());
-                        collect_bindings(&arm.body, vars);
-                        arm_results.push(returned_var(&arm.body).clone());
-                    }
-                    Binding::Join(arm_results)
-                }
-            };
-            vars.bindings.insert(x.name.clone(), def);
-            vars.var_tys.insert(x.name.clone(), x.ty.clone());
-            collect_bindings(k, vars);
-        }
-        RcExpr::Destructure(container, fields, k) => {
-            for (idx, fv) in fields {
-                vars.bindings
-                    .insert(fv.name.clone(), Binding::Field(container.clone(), *idx));
-                vars.var_tys.insert(fv.name.clone(), fv.ty.clone());
-            }
-            collect_bindings(k, vars);
-        }
-        RcExpr::Retain(_, _, _, k) | RcExpr::Release(_, _, _, k) | RcExpr::Eval(_, k) => {
-            collect_bindings(k, vars)
-        }
-    }
-}
-
-/// The variable an expression returns: the one that its final `Ret` names. Every construct of the RC
-/// IR has a single continuation, so a body has exactly one such `Ret` (a match's arms each have their
-/// own, which return the arm's value to the match binding).
-fn returned_var(node: &RcExprNode) -> &RcVar {
-    grow_stack(|| match node.expr.as_ref() {
-        RcExpr::Ret(v) => v,
-        RcExpr::Let(_, _, k)
-        | RcExpr::Destructure(_, _, k)
-        | RcExpr::Retain(_, _, _, k)
-        | RcExpr::Release(_, _, _, k)
-        | RcExpr::Eval(_, k) => returned_var(k),
-    })
-}
-
-/// Where the object at a leaf comes from.
-#[derive(Debug, PartialEq, Eq)]
-enum Origin {
-    /// The leaf denotes exactly this object.
-    Exactly(VarPath),
-    /// The leaf denotes one of `candidates`, chosen by the path taken. `identity` names the match
-    /// binding that joins them: every alias chain through that binding agrees on the name, so it is
-    /// the name to use where one name for the value is required.
-    Join {
-        identity: VarPath,
-        candidates: Set<VarPath>,
-    },
-}
-
-impl Origin {
-    /// The one name for the value, for a reader that pairs operations on it — reference-count
-    /// cancellation pairs a retain with the release that un-bumps it, and only a single identity can
-    /// decide that. Two leaves with the same identity hold the same reference.
-    fn identity(&self) -> &VarPath {
-        match self {
-            Origin::Exactly(p) => p,
-            Origin::Join { identity, .. } => identity,
-        }
-    }
-
-    /// Every object the leaf may denote, for a reader whose answer has to hold on all paths.
-    fn candidates(&self) -> Vec<&VarPath> {
-        match self {
-            Origin::Exactly(p) => vec![p],
-            Origin::Join { candidates, .. } => candidates.iter().collect(),
-        }
-    }
-
-    /// Every object an operation on the leaf acts on: the reference the leaf holds, which `identity`
-    /// names, and the object that reference belongs to, which is any of `candidates`.
-    fn acted_on(&self) -> Vec<&VarPath> {
-        let mut out = vec![self.identity()];
-        out.extend(
-            self.candidates()
-                .into_iter()
-                .filter(|p| *p != self.identity()),
-        );
-        out
-    }
-}
-
-/// Where a leaf's object comes from: follow alias edges (move-binds, pure projections, unboxed-union
-/// payloads, catch-all payloads) back to the variable that produced it. The variable is a parameter
-/// when the leaf ultimately comes from an input, and a `Join` when a match forwards several arms'
-/// values to one binding.
-fn origin(vars: &VarTable, type_env: &TypeEnv, var: &FullName, path: &[usize]) -> Origin {
-    grow_stack(|| origin_inner(vars, type_env, var, path))
-}
-
-fn origin_inner(vars: &VarTable, type_env: &TypeEnv, var: &FullName, path: &[usize]) -> Origin {
-    let here = || Origin::Exactly((var.clone(), path.to_vec()));
-    match vars.bindings.get(var) {
-        None | Some(Binding::Param) | Some(Binding::Producer) => here(),
-        Some(Binding::Move(y)) => origin(vars, type_env, &y.name, path),
-        Some(Binding::Join(arm_results)) => {
-            // The arms bind their values to one variable, so a path into it is a path into each of
-            // them. Arms that all reach the same object leave the value exact.
-            let mut candidates = Set::default();
-            for r in arm_results {
-                for p in origin(vars, type_env, &r.name, path).candidates() {
-                    candidates.insert(p.clone());
-                }
-            }
-            match candidates.len() {
-                1 => Origin::Exactly(
-                    candidates
-                        .into_iter()
-                        .next()
-                        .expect("a one-element set has an element"),
-                ),
-                _ => Origin::Join {
-                    identity: (var.clone(), path.to_vec()),
-                    candidates,
-                },
-            }
-        }
-        Some(Binding::Llvm(llvm_gen, args, result_ty)) => {
-            // Constructing an unboxed union lays its payload in place, so the whole union's root is
-            // the payload's root — the construction alias edge, dual to reading a payload out with
-            // `match`. The whole-union path is where this matters: a leaf path descends into the
-            // active variant, which the projection rule below already aliases through `result_prov`.
-            if path.is_empty()
-                && !args.is_empty()
-                && llvm_gen.as_any().is::<InlineLLVMMakeUnionBody>()
-                && !result_ty.is_box(type_env)
-            {
-                return origin(vars, type_env, &args[0].name, &[]);
-            }
-            let arg_tys: Vec<Arc<TypeNode>> = args.iter().map(|a| a.ty.clone()).collect();
-            let decl = llvm_gen.result_prov(result_ty, &arg_tys, type_env);
-            // A result leaf that is a single `Arg(j, p)` is a pure projection of argument `j`'s leaf
-            // `p` — an alias; anything else (a fresh allocation, a boxed-container read, a join of
-            // several sources) is a producer, stopping here. An `Llvm` op is never partially applied,
-            // so a well-formed `result_prov` names only real argument indices (`args[j]` else panics).
-            // A path with no declared leaf is not a projection either: a reference-counting unit
-            // path may name the root of an unboxed union, which is a subtree rather than a leaf.
-            match decl.leaf_origins_at(path).and_then(as_arg_projection) {
-                Some((j, p)) => origin(vars, type_env, &args[j].name, &p),
-                None => here(),
-            }
-        }
-        Some(Binding::Field(container, idx)) => {
-            if container.ty.is_box(type_env) {
-                // Reading a field of a boxed struct retains it: a producer.
-                here()
-            } else {
-                let mut p = vec![*idx];
-                p.extend_from_slice(path);
-                origin(vars, type_env, &container.name, &p)
-            }
-        }
-        Some(Binding::Payload(scrut, variant)) => match variant {
-            // A catch-all binds the whole scrutinee: the same object.
-            None => origin(vars, type_env, &scrut.name, path),
-            // An unboxed union's payload is the scrutinee's variant slot — an alias; a boxed union's
-            // payload is read out (retained) — a producer.
-            Some(k) if !scrut.ty.is_box(type_env) => {
-                let mut p = vec![*k];
-                p.extend_from_slice(path);
-                origin(vars, type_env, &scrut.name, &p)
-            }
-            Some(_) => here(),
-        },
-    }
-}
-
-/// The single `Arg(j, p)` a leaf source consists of, if it is exactly that.
-fn as_arg_projection(ls: &Set<LeafOrigin>) -> Option<(usize, FieldPath)> {
-    if ls.len() != 1 {
-        return None;
-    }
-    match ls.iter().next() {
-        Some(LeafOrigin::Arg(j, p)) => Some((*j, p.clone())),
-        _ => None,
-    }
-}
-
-/// Collect the leaves consumed in a function body, given `own` as the owned parameter leaves that
-/// decide which argument positions consume: an owning argument position, a captured value, or a
-/// returned value. Alias edges are not consumes here — the consume of an alias is attributed to its
-/// `root`. Explicit `Release` nodes are own-then-release drops, not consumes.
-fn collect_consumes(
-    node: &RcExprNode,
-    vars: &VarTable,
-    prog: &RcProgram,
-    own: &Set<VarPath>,
-    type_env: &TypeEnv,
-    out: &mut Vec<VarPath>,
-) {
-    let owns = |p: &RcVar, pi: &FieldPath| own.contains(&(p.name.clone(), pi.clone()));
-    collect_consumes_go(node, vars, prog, type_env, &owns, out);
-}
-
-fn collect_consumes_go<F: Fn(&RcVar, &FieldPath) -> bool>(
-    node: &RcExprNode,
-    vars: &VarTable,
-    prog: &RcProgram,
-    type_env: &TypeEnv,
-    owns: &F,
-    out: &mut Vec<VarPath>,
-) {
-    match node.expr.as_ref() {
-        RcExpr::Ret(x) => push_boxed_leaves(&x.name, &x.ty, type_env, out),
-        RcExpr::Let(x, rhs, k) => {
-            match rhs {
-                RcRhs::Match(_, arms) => {
-                    for arm in arms {
-                        collect_consumes_go(&arm.body, vars, prog, type_env, owns, out);
-                    }
-                }
-                _ => rhs_consumes(rhs, &x.ty, vars, prog, type_env, owns, out),
-            }
-            collect_consumes_go(k, vars, prog, type_env, owns, out);
-        }
-        RcExpr::Destructure(container, fields, k) => {
-            for pi in destructure_consumes(container, fields, type_env) {
-                out.push((container.name.clone(), pi));
-            }
-            collect_consumes_go(k, vars, prog, type_env, owns, out)
-        }
-        RcExpr::Retain(_, _, _, k) | RcExpr::Release(_, _, _, k) | RcExpr::Eval(_, k) => {
-            collect_consumes_go(k, vars, prog, type_env, owns, out)
-        }
-    }
-}
-
-/// The container leaves a `Destructure` consumes. A boxed container is released whole, so every boxed
-/// leaf of it goes; an unboxed container moves each named field's leaves into that field's variable,
-/// an alias whose consume is attributed to the field variable, so only a dropped (unnamed) field's
-/// leaves go. This is the model code generation implements (`ObjectFieldType::get_struct_fields`), and
-/// every reader of the consume model shares it.
-fn destructure_consumes(
-    container: &RcVar,
-    fields: &[(usize, RcVar)],
-    type_env: &TypeEnv,
-) -> Vec<FieldPath> {
-    let leaves = boxed_leaves(&container.ty, type_env);
-    if container.ty.is_box(type_env) {
-        return leaves;
-    }
-    let named: Set<usize> = fields.iter().map(|(i, _)| *i).collect();
-    leaves
-        .into_iter()
-        .filter(|pi| {
-            // A boxed leaf of an unboxed container starts with a field index, so its path is non-empty.
-            let field = pi
-                .first()
-                .expect("a boxed leaf of an unboxed container has a non-empty path");
-            !named.contains(field)
-        })
-        .collect()
-}
-
-/// The leaves an `App`, `Llvm`, or `Closure` right-hand side consumes: an owning argument position
-/// (`owns` decides, for the callee's parameter leaf), a captured value, and the closure callee. A
-/// `Var` move and a `Match` consume nothing here — a move is an alias, and a match's consumes live in
-/// its arms. `result_ty` is the type the right-hand side binds, needed to read an op's passthrough.
-fn rhs_consumes<F: Fn(&RcVar, &FieldPath) -> bool>(
-    rhs: &RcRhs,
-    result_ty: &Arc<TypeNode>,
-    vars: &VarTable,
-    prog: &RcProgram,
-    type_env: &TypeEnv,
-    owns: &F,
-    out: &mut Vec<VarPath>,
-) {
-    match rhs {
-        RcRhs::Var(_) | RcRhs::Match(..) => {}
-        RcRhs::Closure(_, caps) => {
-            for c in caps {
-                push_boxed_leaves(&c.name, &c.ty, type_env, out);
-            }
-        }
-        RcRhs::App(callee, args) => {
-            // Calling a closure consumes it (the callee releases its capture).
-            push_boxed_leaves(&callee.name, &callee.ty, type_env, out);
-            // Each argument at an owning position of the callee is consumed. An unresolved (indirect)
-            // callee owns every position.
-            let callee_params = resolve_callee_params(callee, vars, prog);
-            for (i, a) in args.iter().enumerate() {
-                for pi in boxed_leaves(&a.ty, type_env) {
-                    // `i` ranges over the arguments and `args.len() <= params.len()` (no over-
-                    // application), so `params[i]` is in range.
-                    let owns_pos = match &callee_params {
-                        Some(params) => owns(&params[i], &pi),
-                        None => true,
-                    };
-                    if owns_pos {
-                        out.push((a.name.clone(), pi));
-                    }
-                }
-            }
-        }
-        RcRhs::Llvm(llvm_gen, args) => {
-            let arg_tys: Vec<Arc<TypeNode>> = args.iter().map(|a| a.ty.clone()).collect();
-            let passthrough = passthrough_arg_leaves(&**llvm_gen, result_ty, args, type_env);
-            for (i, a) in args.iter().enumerate() {
-                if llvm_gen.borrows_operand(i, &arg_tys, type_env) {
-                    continue;
-                }
-                for pi in boxed_leaves(&a.ty, type_env) {
-                    // An argument leaf that the op passes through to its result is not consumed;
-                    // anything else at an owning position is moved into the op.
-                    if !passthrough.contains(&(i, pi.clone())) {
-                        out.push((a.name.clone(), pi));
-                    }
-                }
-            }
-        }
-    }
-}
-
-/// The parameters of a directly-called callee: a closure value built in this function, or a
-/// top-level function referenced by name. `None` for an indirect call (an owning, all-`Own` ABI).
-fn resolve_callee_params<'a>(
-    callee: &RcVar,
-    vars: &VarTable,
-    prog: &'a RcProgram,
-) -> Option<&'a [RcVar]> {
-    let fref = vars
-        .closure_targets
-        .get(&callee.name)
-        .cloned()
-        .or_else(|| {
-            let fref = FuncRef {
-                name: callee.name.clone(),
-            };
-            prog.funcs.contains_key(&fref).then_some(fref)
-        })?;
-    // A closure target is registered as a function when it is lifted, and the other way in resolved
-    // the name against the program already, so the callee is there either way. Reporting it absent
-    // here would read as an indirect call and quietly give up the borrow optimization.
-    let func = prog.funcs.get(&fref).unwrap_or_else(|| {
-        unreachable!(
-            "callee `{}` is not a function of the program",
-            fref.name.to_string()
-        )
-    });
-    Some(func.params.as_slice())
-}
-
-/// The `(arg index, leaf path)` pairs an LLVM op passes through unchanged to its result — the pure
-/// projections `as_arg_projection` reads out of `result_prov`.
-///
-/// Dropping an argument leaf's consume is sound exactly when the result aliases it, so this shares
-/// `as_arg_projection` with `root`: a leaf that joins an argument with another source aliases nothing and
-/// keeps its consume, and one whose sole source is `Arg` does both.
-fn passthrough_arg_leaves(
-    llvm_gen: &dyn LLVMGen,
-    result_ty: &Arc<TypeNode>,
-    args: &[RcVar],
-    type_env: &TypeEnv,
-) -> Set<(usize, FieldPath)> {
-    let arg_tys: Vec<Arc<TypeNode>> = args.iter().map(|a| a.ty.clone()).collect();
-    let decl = llvm_gen.result_prov(result_ty, &arg_tys, type_env);
-    decl.leaves().filter_map(as_arg_projection).collect()
-}
-
-/// Push every boxed leaf of a value onto `out`.
-fn push_boxed_leaves(
-    var: &FullName,
-    ty: &Arc<TypeNode>,
-    type_env: &TypeEnv,
-    out: &mut Vec<VarPath>,
-) {
-    for p in boxed_leaves(ty, type_env) {
-        out.push((var.clone(), p));
-    }
-}
-
-/// The paths of every boxed leaf of a type: the whole value if boxed, the capture of a closure, or
-/// each boxed leaf of an unboxed aggregate.
-fn boxed_leaves(ty: &Arc<TypeNode>, type_env: &TypeEnv) -> Vec<FieldPath> {
-    boxed_leaf_paths(ty, type_env)
-}
-
-// --- reference-counting units ---
-
-/// The reference-counting units of a value's type: the capture of a closure, or each unit root
-/// (`is_rc_unit_root`) — a boxed value, an unboxed union, or a punched array — reached by descending
-/// its unboxed structs/tuples. Unlike `boxed_leaves`, it stops at a unit root rather than expanding it
-/// into the inner boxed leaves (e.g. an unboxed union is one unit, since only its active variant is
-/// live and a refcount operation must dispatch on the tag rather than name a variant's leaf).
-pub fn rc_units(ty: &Arc<TypeNode>, type_env: &TypeEnv) -> Vec<FieldPath> {
-    let mut out = vec![];
-    rc_units_go(ty, type_env, &mut vec![], &mut out);
-    out
-}
-
-fn rc_units_go(
-    ty: &Arc<TypeNode>,
-    type_env: &TypeEnv,
-    path: &mut FieldPath,
-    out: &mut Vec<FieldPath>,
-) {
-    if ty.is_fully_unboxed(type_env) {
-        return;
-    }
-    if ty.is_closure() {
-        path.push(CLOSURE_CAPTURE_IDX as usize);
-        out.push(path.clone());
-        path.pop();
-        return;
-    }
-    if ty.is_rc_unit_root(type_env) {
-        out.push(path.clone());
-        return;
-    }
-    let fields = ty.fields(type_env);
-    for (i, fty) in ty.field_types(type_env).iter().enumerate() {
-        // A punched struct field is a hole (its value has moved out): the whole-value traversal skips
-        // it, so it names no unit.
-        if fields[i].is_punched {
-            continue;
-        }
-        path.push(i);
-        rc_units_go(fty, type_env, path, out);
-        path.pop();
-    }
-}
-
-/// Truncate a leaf path to its reference-counting unit: the path down to the first unit root
-/// (`is_rc_unit_root`) it reaches — an unboxed union or a punched array, whose subtree is one unit.
-/// Paths that stay within unboxed structs are unchanged.
-fn truncate_to_unit(ty: &Arc<TypeNode>, path: &[usize], type_env: &TypeEnv) -> FieldPath {
-    let mut out = vec![];
-    let mut cur = ty.clone();
-    for &idx in path {
-        if cur.is_closure() {
-            // The only path into a closure names its capture, which is a single unit.
-            out.push(idx);
-            break;
-        }
-        if cur.is_rc_unit_root(type_env) {
-            // A boxed value, an unboxed union, or a punched array is one unit; a leaf below it (a
-            // boxed leaf under a union variant, or the punched array's inner array) keys to its root.
-            break;
-        }
-        // Here `cur` is an unboxed struct/tuple, so a well-formed unit/root path index is in range.
-        let fields = cur.field_types(type_env);
-        assert!(
-            idx < fields.len(),
-            "truncate_to_unit: path index {} out of range ({} fields)",
-            idx,
-            fields.len()
-        );
-        out.push(idx);
-        cur = fields[idx].clone();
-    }
-    out
 }
 
 // --- borrow-ification ---
@@ -762,23 +223,6 @@ pub fn borrow_ify(prog: &RcProgram, type_env: &TypeEnv) -> RcProgram {
         globals,
         entry: prog.entry.clone(),
     }
-}
-
-/// The owned parameter/capture units of every function: each version's units minus the ones it
-/// borrows (`RcFunc::borrowed_units`, the annotation borrow-ification writes).
-fn all_owned_units(prog: &RcProgram, type_env: &TypeEnv) -> Set<VarPath> {
-    let mut owned = Set::default();
-    for func in prog.funcs.values() {
-        for p in func.params.iter().chain(func.capture.iter()) {
-            for unit in rc_units(&p.ty, type_env) {
-                let leaf = (p.name.clone(), unit);
-                if !func.borrowed_units.contains(&leaf) {
-                    owned.insert(leaf);
-                }
-            }
-        }
-    }
-    owned
 }
 
 /// Each parameter/capture variable's ownership shape, derived from the functions' ownership
@@ -1238,43 +682,6 @@ fn rhs_uses(name: &FullName, rhs: &RcRhs) -> bool {
     }
 }
 
-/// The reference-counting units under a path of a value's type: the units of the subtree the path
-/// names, or the path itself when it already names a unit (a boxed value, a union, or a leaf).
-fn units_under(ty: &Arc<TypeNode>, path: &FieldPath, type_env: &TypeEnv) -> Vec<FieldPath> {
-    match subtree_type(ty, path, type_env) {
-        Some(sty) => rc_units(&sty, type_env)
-            .into_iter()
-            .map(|u| {
-                let mut p = path.clone();
-                p.extend(u);
-                p
-            })
-            .collect(),
-        None => vec![path.clone()],
-    }
-}
-
-/// The type of the subtree a path names, descending only unboxed structs; `None` once the path
-/// reaches a closure, a unit root (`is_rc_unit_root`), or a fully-unboxed leaf.
-fn subtree_type(ty: &Arc<TypeNode>, path: &FieldPath, type_env: &TypeEnv) -> Option<Arc<TypeNode>> {
-    let mut cur = ty.clone();
-    for &idx in path {
-        if cur.is_closure() || cur.is_rc_unit_root(type_env) || cur.is_fully_unboxed(type_env) {
-            return None;
-        }
-        // Here `cur` is an unboxed struct/tuple, so a well-formed unit/root path index is in range.
-        let fields = cur.field_types(type_env);
-        assert!(
-            idx < fields.len(),
-            "subtree_type: path index {} out of range ({} fields)",
-            idx,
-            fields.len()
-        );
-        cur = fields[idx].clone();
-    }
-    Some(cur)
-}
-
 // --- unit normalization ---
 
 /// Decompose every `Retain`/`Release` into one node per reference-counting unit its path covers, so
@@ -1727,165 +1134,5 @@ fn drop_nodes_inner(node: &RcExprNode, to_delete: &Set<NodeId>) -> RcExprNode {
             &node.source,
         ),
         RcExpr::Ret(v) => expr_node(RcExpr::Ret(v.clone()), &node.source),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::fixstd::builtin::make_i64_ty;
-    use crate::rc_ir::provenance::Provenance;
-
-    fn sources(srcs: Vec<LeafOrigin>) -> Set<LeafOrigin> {
-        srcs.into_iter().collect()
-    }
-
-    #[test]
-    fn a_lone_arg_is_a_projection() {
-        let ls = Provenance::leaf(LeafOrigin::Arg(1, vec![0]));
-        assert_eq!(as_arg_projection(&ls), Some((1, vec![0])));
-    }
-
-    #[test]
-    fn an_arg_joined_with_another_source_is_not_a_projection() {
-        // The result is the argument or a new value, so it aliases neither: the op consumes the
-        // argument, and `root` stops at the op. Reading such a leaf as a projection would drop the
-        // consume without the alias, releasing one object twice.
-        let ls = sources(vec![LeafOrigin::Fresh, LeafOrigin::Arg(0, vec![])]);
-        assert_eq!(as_arg_projection(&ls), None);
-    }
-
-    #[test]
-    fn one_of_two_args_is_not_a_projection() {
-        // Neither argument can be named as the alias, whichever of the two the set yields first.
-        let ls = sources(vec![LeafOrigin::Arg(0, vec![]), LeafOrigin::Arg(1, vec![])]);
-        assert_eq!(as_arg_projection(&ls), None);
-    }
-
-    #[test]
-    fn a_produced_leaf_is_not_a_projection() {
-        assert_eq!(
-            as_arg_projection(&Provenance::leaf(LeafOrigin::Fresh)),
-            None
-        );
-        assert_eq!(
-            as_arg_projection(&Provenance::leaf(LeafOrigin::Unknown)),
-            None
-        );
-    }
-
-    #[test]
-    fn a_bottom_leaf_is_not_a_projection() {
-        // `_undefined_internal` aborts, so its result has no source at all.
-        assert_eq!(as_arg_projection(&sources(vec![])), None);
-    }
-
-    fn var(name: &str) -> RcVar {
-        RcVar {
-            name: FullName::local(name),
-            ty: make_i64_ty(),
-            source: None,
-            debug_name: None,
-            skip_null_check: false,
-        }
-    }
-
-    /// A table of the given bindings, with every named variable also a known local.
-    fn table(bindings: Vec<(&str, Binding)>) -> VarTable {
-        let mut vars = VarTable::empty();
-        for (name, b) in bindings {
-            let v = var(name);
-            vars.bindings.insert(v.name.clone(), b);
-            vars.var_tys.insert(v.name, v.ty);
-        }
-        vars
-    }
-
-    fn at(name: &str) -> VarPath {
-        (FullName::local(name), vec![])
-    }
-
-    fn origin_of(vars: &VarTable, name: &str) -> Origin {
-        origin(vars, &TypeEnv::default(), &FullName::local(name), &[])
-    }
-
-    #[test]
-    fn a_producer_is_exactly_itself() {
-        let vars = table(vec![("p", Binding::Producer)]);
-        assert_eq!(origin_of(&vars, "p"), Origin::Exactly(at("p")));
-    }
-
-    #[test]
-    fn a_move_bind_is_the_moved_variable() {
-        let vars = table(vec![
-            ("p", Binding::Producer),
-            ("m", Binding::Move(var("p"))),
-        ]);
-        assert_eq!(origin_of(&vars, "m"), Origin::Exactly(at("p")));
-    }
-
-    #[test]
-    fn a_match_binding_may_be_any_arm_result() {
-        // The two arms produce different objects, so the binding is one of them and the join itself
-        // is the name every alias chain through it agrees on.
-        let vars = table(vec![
-            ("p", Binding::Producer),
-            ("q", Binding::Producer),
-            ("m", Binding::Join(vec![var("p"), var("q")])),
-        ]);
-        let o = origin_of(&vars, "m");
-        assert_eq!(o.identity(), &at("m"));
-        assert_eq!(
-            o.candidates()
-                .into_iter()
-                .cloned()
-                .collect::<Set<VarPath>>(),
-            vec![at("p"), at("q")].into_iter().collect::<Set<VarPath>>()
-        );
-    }
-
-    #[test]
-    fn a_match_binding_whose_arms_agree_is_exact() {
-        // Both arms reach `p`, one of them through a move-bind, so the binding is `p` on every path.
-        let vars = table(vec![
-            ("p", Binding::Producer),
-            ("m1", Binding::Move(var("p"))),
-            ("m", Binding::Join(vec![var("p"), var("m1")])),
-        ]);
-        assert_eq!(origin_of(&vars, "m"), Origin::Exactly(at("p")));
-    }
-
-    #[test]
-    fn a_move_of_a_match_binding_keeps_the_joins_name() {
-        // The identity has to survive an alias chain, or a retain of the binding and a release of the
-        // moved-to variable would key differently and never pair.
-        let vars = table(vec![
-            ("p", Binding::Producer),
-            ("q", Binding::Producer),
-            ("m", Binding::Join(vec![var("p"), var("q")])),
-            ("n", Binding::Move(var("m"))),
-        ]);
-        assert_eq!(origin_of(&vars, "n").identity(), &at("m"));
-    }
-
-    #[test]
-    fn a_join_of_joins_may_be_any_of_their_results() {
-        let vars = table(vec![
-            ("p", Binding::Producer),
-            ("q", Binding::Producer),
-            ("r", Binding::Producer),
-            ("inner", Binding::Join(vec![var("p"), var("q")])),
-            ("m", Binding::Join(vec![var("inner"), var("r")])),
-        ]);
-        assert_eq!(
-            origin_of(&vars, "m")
-                .candidates()
-                .into_iter()
-                .cloned()
-                .collect::<Set<VarPath>>(),
-            vec![at("p"), at("q"), at("r")]
-                .into_iter()
-                .collect::<Set<VarPath>>()
-        );
     }
 }
